Move percent attribute functionality into own object
Except for the attribute store functions.
